move format control to models and write parametrized test
