+++ src/pip/_internal/commands/freeze.py
@@ -4,7 +4,7 @@ import sys
 
 from pip._internal.cache import WheelCache
 from pip._internal.cli.base_command import Command
-from pip._internal.format_control import FormatControl
+from pip._internal.models.format_control import FormatControl
 from pip._internal.operations.freeze import freeze
 from pip._internal.utils.compat import stdlib_pkgs
